refactor apollo and express server files
